remove entity column duplicate
