+++ src/main/java/codesquad/domain/BaseEntity.java
@@ -1,0 +1,43 @@
+package codesquad.domain;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+@Getter
+@Setter
+@EqualsAndHashCode(exclude = {"createdDate", "modifiedDate"})
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public class BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @CreatedDate
+    private LocalDateTime createdDate;
+
+    @LastModifiedDate
+    private LocalDateTime modifiedDate;
+
+    public boolean isMatch(Long id) {
+        if (id == null) {
+            return false;
+        }
+        return this.id.equals(id);
+    }
+
+    public String getFormattedDate() {
+        if (modifiedDate == null) {
+            return "";
+        }
+        return modifiedDate.format(DateTimeFormatter.ofPattern("yyyy.MM.dd HH:mm:ss"));
+    }
+}
