--- src/main/java/codesquad/domain/TimeEntity.java
@@ -1,26 +1,0 @@
-package codesquad.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class TimeEntity {
-
-    @CreatedDate
-    @JsonIgnore
-    private LocalDateTime createdDate;
-
-    @LastModifiedDate
-    @JsonProperty
-    private LocalDateTime modifiedDate;
-}
