+++ src/main/java/codesquad/domain/answer/Answer.java
@@ -1,6 +1,6 @@
 package codesquad.domain.answer;
 
-import codesquad.domain.TimeEntity;
+import codesquad.domain.BaseEntity;
 import codesquad.domain.exception.ForbiddenRequestException;
 import codesquad.domain.exception.UnAuthorizedException;
 import codesquad.domain.question.Question;
@@ -17,11 +17,7 @@ import java.util.Optional;
 @Setter
 @ToString
 @Entity
-public class Answer extends TimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+public class Answer extends BaseEntity {
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_answer_user"))
     private User user;
