+++ src/main/resources/templates/index.html
@@ -21,7 +21,7 @@
                           </strong>
                           <div class="auth-info">
                               <i class="icon-add-comment"></i>
-                              <span class="time">{{modifiedDate}}</span>
+                              <span class="time">{{formattedDate}}</span>
                               <a href="/users/{{user.id}}" class="author">{{user.name}}</a>
                           </div>
                           <div class="reply" title="댓글">
